pyproject.toml
[tool.poetry]
name = "python-auth"
version = "0.1.0"
description = "Python authentication microservice"
authors = []
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
pydantic = "*"
pydantic-settings = "*"
sqlalchemy = {extras = ["asyncio"], version = "*"}
asyncpg = "*"
alembic = "*"
boto3 = "*"
python-jose = {extras = ["cryptography"], version = "*"}
structlog = "*"
greenlet = "^3.3.2"
httpx = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
ruff = "*"
mypy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
